user  dashbaord fix

+++ src/main/java/com/example/carrent/dtos/car/CarCreateDto.java
@@ -2,6 +2,7 @@ package com.example.carrent.dtos.car;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -9,7 +10,7 @@ public class CarCreateDto {
     private String brand;
     private String model;
     private String imageUrl;
-    private Double dailyPrice;
+    private BigDecimal dailyPrice;
     private Integer doorCount;
     private Integer passengerCount;
     private Integer luggageCapacity;
